requirements.txt
boto3==1.35.90
botocore==1.35.90
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rocksdict==0.3.25
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
